Refactor to traits
