+++ app/Models/Concerns/CanBeRetired.php
@@ -1,0 +1,77 @@
+<?php
+
+namespace App\Models\Concerns;
+
+use App\Models\Retirement;
+
+trait CanBeRetired
+{
+    /**
+     * Get the retirements of the model.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\MorphMany
+     */
+    public function retirements()
+    {
+        return $this->morphMany(Retirement::class, 'retiree');
+    }
+
+    /**
+     * Get the current retirement of the model.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\MorphOne
+     */
+    public function retirement()
+    {
+        return $this->morphOne(Retirement::class, 'retiree')->whereNull('ended_at');
+    }
+
+    /**
+     * Determine if a model is retired.
+     *
+     * @return bool
+     */
+    public function getIsRetiredAttribute()
+    {
+        return $this->retirement()->exists();
+    }
+
+    /**
+     * Scope a query to only include retired models.
+     *
+     * @param  \Illuminate\Database\Eloquent\Builder  $query
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public function scopeRetired($query)
+    {
+        return $query->whereHas('retirement');
+    }
+
+    /**
+     * Retire a model.
+     *
+     * @return \App\Models\Retirement
+     */
+    public function retire()
+    {
+        if ($this->is_suspended) {
+            $this->reinstate();
+        }
+
+        if ($this->is_injured) {
+            $this->recover();
+        }
+
+        $this->retirements()->create(['started_at' => now()]);
+    }
+
+    /**
+     * Unretire a model.
+     *
+     * @return bool
+     */
+    public function unretire()
+    {
+        return $this->retirement()->update(['ended_at' => now()]);
+    }
+}
