+++ app/Models/Referee.php
@@ -53,7 +53,11 @@ use Illuminate\Database\Eloquent\SoftDeletes;
  */
 class Referee extends Model
 {
-    use SoftDeletes;
+    use SoftDeletes,
+        Concerns\CanBeSuspended,
+        Concerns\CanBeInjured,
+        Concerns\CanBeRetired,
+        Concerns\CanBeEmployed;
 
     /**
      * The attributes that aren't mass assignable.
@@ -61,86 +65,6 @@ class Referee extends Model
      * @var array
      */
     protected $guarded = [];
-
-    /**
-     * Get all of the employments of the referee.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphMany
-     */
-    public function employments()
-    {
-        return $this->morphMany(Employment::class, 'employable')->whereNull('ended_at');
-    }
-
-    /**
-     * Get the current employment of the referee.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphOne
-     */
-    public function employment()
-    {
-        return $this->morphOne(Employment::class, 'employable')->whereNull('ended_at');
-    }
-
-    /**
-     * Get the retirements of the referee.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphMany
-     */
-    public function retirements()
-    {
-        return $this->morphMany(Retirement::class, 'retiree');
-    }
-
-    /**
-     * Get the current retirement of the referee.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphOne
-     */
-    public function retirement()
-    {
-        return $this->morphOne(Retirement::class, 'retiree')->whereNull('ended_at');
-    }
-
-    /**
-     * Get the suspensions of the referee.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphMany
-     */
-    public function suspensions()
-    {
-        return $this->morphMany(Suspension::class, 'suspendable');
-    }
-
-    /**
-     * Get the current suspension of the referee.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphOne
-     */
-    public function suspension()
-    {
-        return $this->morphOne(Suspension::class, 'suspendable')->whereNull('ended_at');
-    }
-
-    /**
-     * Get the injuries of the referee.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphMany
-     */
-    public function injuries()
-    {
-        return $this->morphMany(Injury::class, 'injurable');
-    }
-
-    /**
-     * Get the current injury of the referee.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphOne
-     */
-    public function injury()
-    {
-        return $this->morphOne(Injury::class, 'injurable')->whereNull('ended_at');
-    }
 
     /**
      * Determine the status of the referee.
@@ -180,46 +104,6 @@ class Referee extends Model
     }
 
     /**
-     * Determine if a referee is hired.
-     *
-     * @return bool
-     */
-    public function getIsEmployedAttribute()
-    {
-        return $this->employments()->where('started_at', '<=', now())->whereNull('ended_at')->exists();
-    }
-
-    /**
-     * Determine if a referee is retired.
-     *
-     * @return bool
-     */
-    public function getIsRetiredAttribute()
-    {
-        return $this->retirements()->whereNull('ended_at')->exists();
-    }
-
-    /**
-     * Determine if a referee is suspended.
-     *
-     * @return bool
-     */
-    public function getIsSuspendedAttribute()
-    {
-        return $this->suspensions()->whereNull('ended_at')->exists();
-    }
-
-    /**
-     * Determine if a referee is injured.
-     *
-     * @return bool
-     */
-    public function getIsInjuredAttribute()
-    {
-        return $this->injuries()->whereNull('ended_at')->exists();
-    }
-
-    /**
      * Get the full name of the referee.
      *
      * @return string
@@ -245,135 +129,6 @@ class Referee extends Model
         })->whereDoesntHave('suspensions', function (Builder $query) {
             $query->whereNull('ended_at');
         });
-    }
-
-    /**
-     * Scope a query to only include pending introduced referees.
-     *
-     * @param  \Illuminate\Database\Eloquent\Builder $query
-     */
-    public function scopePendingIntroduction($query)
-    {
-        return $query->whereHas('employments', function (Builder $query) {
-            $query->whereNull('started_at')->orWhere('started_at', '>', now());
-        });
-    }
-
-    /**
-     * Scope a query to only include retired referees.
-     *
-     * @param  \Illuminate\Database\Eloquent\Builder  $query
-     * @return \Illuminate\Database\Eloquent\Builder
-     */
-    public function scopeRetired($query)
-    {
-        return $query->whereHas('retirements', function ($query) {
-            $query->whereNull('ended_at');
-        });
-    }
-
-    /**
-     * Scope a query to only include suspended referees.
-     *
-     * @param  \Illuminate\Database\Eloquent\Builder  $query
-     * @return \Illuminate\Database\Eloquent\Builder
-     */
-    public function scopeSuspended($query)
-    {
-        return $query->whereHas('suspensions', function ($query) {
-            $query->whereNull('ended_at');
-        });
-    }
-
-    /**
-     * Scope a query to only include injured referees.
-     *
-     * @param  \Illuminate\Database\Eloquent\Builder  $query
-     * @return \Illuminate\Database\Eloquent\Builder
-     */
-    public function scopeInjured($query)
-    {
-        return $query->whereHas('injuries', function ($query) {
-            $query->whereNull('ended_at');
-        });
-    }
-
-    /**
-     * Activate a referee.
-     *
-     * @return bool
-     */
-    public function activate()
-    {
-        return $this->employments()->latest()->first()->update(['started_at' => now()]);
-    }
-
-    /**
-     * Retire a referee.
-     *
-     * @return \App\Models\Retirement
-     */
-    public function retire()
-    {
-        if ($this->is_suspended) {
-            $this->reinstate();
-        }
-
-        if ($this->is_injured) {
-            $this->recover();
-        }
-
-        $this->retirements()->create(['started_at' => now()]);
-    }
-
-    /**
-     * Unretire a referee.
-     *
-     * @return bool
-     */
-    public function unretire()
-    {
-        return $this->retirement()->update(['ended_at' => now()]);
-    }
-
-    /**
-     * Injure a referee.
-     *
-     * @return \App\Models\Injury
-     */
-    public function injure()
-    {
-        $this->injuries()->create(['started_at' => now()]);
-    }
-
-    /**
-     * Recover a referee.
-     *
-     * @return bool
-     */
-    public function recover()
-    {
-        $this->injury()->update(['ended_at' => now()]);
-    }
-
-    /**
-     * Suspend a referee.
-     *
-     * @return \App\Models\Suspension
-     */
-    public function suspend()
-    {
-        $this->suspensions()->create(['started_at' => now()]);
-    }
-
-    /**
-     * Reinstate a referee.
-     *
-     * @return bool
-     */
-    public function reinstate()
-    {
-        $this->suspension()->update(['ended_at' => now()]);
     }
 
     /**
