+++ app/Models/Wrestler.php
@@ -71,7 +71,12 @@ use App\Eloquent\Concerns\HasCustomRelationships;
  */
 class Wrestler extends Model
 {
-    use SoftDeletes, HasCustomRelationships;
+    use SoftDeletes,
+        HasCustomRelationships,
+        Concerns\CanBeSuspended,
+        Concerns\CanBeInjured,
+        Concerns\CanBeRetired,
+        Concerns\CanBeEmployed;
 
 
     /**
@@ -131,86 +136,6 @@ class Wrestler extends Model
     public function stable()
     {
         // return $this->morphToMany(Stable::class, 'member')->where('is_active', true);
-    }
-
-    /**
-     * Get the retirements of the wrestler.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphMany
-     */
-    public function retirements()
-    {
-        return $this->morphMany(Retirement::class, 'retiree');
-    }
-
-    /**
-     * Get the current retirement of the wrestler.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphOne
-     */
-    public function retirement()
-    {
-        return $this->morphOne(Retirement::class, 'retiree')->whereNull('ended_at');
-    }
-
-    /**
-     * Get the suspensions of the wrestler.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphMany
-     */
-    public function suspensions()
-    {
-        return $this->morphMany(Suspension::class, 'suspendable');
-    }
-
-    /**
-     * Get the current suspension of the wrestler.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphOne
-     */
-    public function suspension()
-    {
-        return $this->morphOne(Suspension::class, 'suspendable')->whereNull('ended_at');
-    }
-
-    /**
-     * Get the injuries of the wrestler.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphMany
-     */
-    public function injuries()
-    {
-        return $this->morphMany(Injury::class, 'injurable');
-    }
-
-    /**
-     * Get the current injury of the wrestler.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphOne
-     */
-    public function injury()
-    {
-        return $this->morphOne(Injury::class, 'injurable')->whereNull('ended_at');
-    }
-
-    /**
-     * Get all of the employments of the wrestler.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphMany
-     */
-    public function employments()
-    {
-        return $this->morphMany(Employment::class, 'employable')->whereNull('ended_at');
-    }
-
-    /**
-     * Get the current employment of the wrestler.
-     *
-     * @return \Illuminate\Database\Eloquent\Relations\MorphOne
-     */
-    public function employment()
-    {
-        return $this->morphOne(Employment::class, 'employable')->whereNull('ended_at');
     }
 
     /**
@@ -274,47 +199,6 @@ class Wrestler extends Model
     }
 
     /**
-     * Determine if a wrestler is hired.
-     *
-     * @return bool
-     */
-    public function getIsEmployedAttribute()
-    {
-        return $this->employments()->where('started_at', '<=', now())->whereNull('ended_at')->exists();
-    }
-
-    /**
-     * Determine if a wrestler is retired.
-     *
-     * @return bool
-     */
-    public function getIsRetiredAttribute()
-    {
-        return $this->retirements()->whereNull('ended_at')->exists();
-    }
-
-    /**
-     * Determine if a wrestler is suspended.
-     *
-     * @return bool
-     */
-    public function getIsSuspendedAttribute()
-    {
-        return $this->suspensions()->whereNull('ended_at')->exists();
-    }
-
-    /**
-     * Determine if a wrestler is injured.
-     *
-     * @return bool
-     */
-    public function getIsInjuredAttribute()
-    {
-        return $this->injuries()->whereNull('ended_at')->exists();
-    }
-
-
-    /**
      * Return the wrestler's height in feet.
      *
      * @return string
@@ -350,135 +234,6 @@ class Wrestler extends Model
         })->whereDoesntHave('suspensions', function (Builder $query) {
             $query->whereNull('ended_at');
         });
-    }
-
-    /**
-     * Scope a query to only include wrestlers that haven't started.
-     *
-     * @param  \Illuminate\Database\Eloquent\Builder $query
-     */
-    public function scopePendingIntroduction($query)
-    {
-        return $query->whereHas('employments', function (Builder $query) {
-            $query->where('started_at', '>', now());
-        })->orWhereDoesntHave('employment');
-    }
-
-    /**
-     * Scope a query to only include retired wrestlers.
-     *
-     * @param  \Illuminate\Database\Eloquent\Builder  $query
-     * @return \Illuminate\Database\Eloquent\Builder
-     */
-    public function scopeRetired($query)
-    {
-        return $query->whereHas('retirements', function ($query) {
-            $query->whereNull('ended_at');
-        });
-    }
-
-    /**
-     * Scope a query to only include suspended wrestlers.
-     *
-     * @param  \Illuminate\Database\Eloquent\Builder  $query
-     * @return \Illuminate\Database\Eloquent\Builder
-     */
-    public function scopeSuspended($query)
-    {
-        return $query->whereHas('suspensions', function ($query) {
-            $query->whereNull('ended_at');
-        });
-    }
-
-    /**
-     * Scope a query to only include injured wrestlers.
-     *
-     * @param  \Illuminate\Database\Eloquent\Builder  $query
-     * @return \Illuminate\Database\Eloquent\Builder
-     */
-    public function scopeInjured($query)
-    {
-        return $query->whereHas('injuries', function ($query) {
-            $query->whereNull('ended_at');
-        });
-    }
-
-    /**
-     * Activate a wrestler.
-     *
-     * @return bool
-     */
-    public function activate()
-    {
-        return $this->employments()->latest()->first()->update(['started_at' => now()]);
-    }
-
-    /**
-     * Retire a wrestler.
-     *
-     * @return \App\Models\Retirement
-     */
-    public function retire()
-    {
-        if ($this->is_suspended) {
-            $this->reinstate();
-        }
-
-        if ($this->is_injured) {
-            $this->recover();
-        }
-
-        $this->retirements()->create(['started_at' => now()]);
-    }
-
-    /**
-     * Unretire a wrestler.
-     *
-     * @return bool
-     */
-    public function unretire()
-    {
-        return $this->retirement()->update(['ended_at' => now()]);
-    }
-
-    /**
-     * Suspend a wrestler.
-     *
-     * @return \App\Models\Suspension
-     */
-    public function suspend()
-    {
-        $this->suspensions()->create(['started_at' => now()]);
-    }
-
-    /**
-     * Reinstate a wrestler.
-     *
-     * @return bool
-     */
-    public function reinstate()
-    {
-        $this->suspension()->update(['ended_at' => now()]);
-    }
-
-    /**
-     * Injure a wrestler.
-     *
-     * @return \App\Models\Injury
-     */
-    public function injure()
-    {
-        $this->injuries()->create(['started_at' => now()]);
-    }
-
-    /**
-     * Recover a wrestler.
-     *
-     * @return bool
-     */
-    public function recover()
-    {
-        $this->injury()->update(['ended_at' => now()]);
     }
 
     /**
